requirements.txt
setuptools<58.0.0
cryptography==3.4.8
ccxt
pandas
pandas-ta
python-telegram-bot